Wrapper: * Started using ioctl macros in SDL platform. * IDL: combined file copies. * Better support for double and float in ioctl:s.
git-svn-id: svn://localhost/public/MoSync/branches/native-wrapper@1146 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ intlibs/idl-common/parse.cpp
@@ -179,7 +179,7 @@ static Ioctl parseIoctl(const vector<string>& ixs, Interface& inf)
 	{	//add ioctl function to interface
 		Function f;
 		f.name = ioctl.name;
-		f.returnType = "int";
+		f.returnType = "longlong";
 		f.comment = getComment();
 		f.number = sNextFuncNum++;
 		Argument a;
